requirements.txt
fastapi
uvicorn[standard]
google-generativeai
firebase-admin
google-cloud-aiplatform
google-api-python-client
pypdf
python-dotenv
python-multipart